+++ packages/react-ui/__tests__/widgets/testUtils.js
@@ -68,6 +68,7 @@ export const mockEcharts = {
       })),
       showLoading: jest.fn(),
       on: jest.fn(),
+      off: jest.fn(),
       getZr: jest.fn()
     }));
   },